Kedro/ds-project/requirements.txt
absl-py==1.3.0
anyconfig==0.10.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astunparse==1.6.3
atomicwrites==1.4.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.12.0
bleach==5.0.1
build==0.9.0
cachetools==4.2.4
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cookiecutter==2.1.1
coverage==7.0.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dynaconf==3.1.11
entrypoints==0.4
fastjsonschema==2.16.2
flake8==4.0.1
flatbuffers==22.12.6
fsspec==2022.11.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
imageio==2.23.0
importlib-metadata==5.2.0
importlib-resources==5.10.1
iniconfig==1.1.1
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.11.4
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
json5==0.9.10
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.4
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.15.2
kedro==0.18.4
kedro-telemetry==0.2.3
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.4
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nbstripout==0.6.1
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pep517==0.13.0
pickleshare==0.7.5
Pillow==9.3.0
pip-tools==6.12.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.13.0
pyrsistent==0.19.2
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==1.13.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytoolconfig==1.2.4
pytz==2022.7
PyWavelets==1.3.0
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rope==1.5.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminado==0.17.1
text-unidecode==1.3
tifffile==2021.11.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toposort==1.7
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
traitlets==5.8.0
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wincertstore==0.2
wrapt==1.14.1
zipp==3.11.0
